Messages constol info and everything else.

+++ app/controllers/BankController.php
@@ -23,31 +23,73 @@ class BankController {
         return App::view('creatNewAcount');
     }
 
+    public function add($id) {
+        $account = $data = $this->get()->show((int)$id);
+         return App::view('add', ['account'=> $account]);
+    }
+
     public function lsit() {
-     
         $data = $this->get()->showAll();
-        
         return App::view('list', ['accounts' => $data]);
     }
 
     public function newAccount() {
     
-        $array = [
-        'id' => rand(1000000000, 9999999999),
-        'name' => $_POST['firstName'],
-        'lastName' =>$_POST['lastName'],
-        'personCode' => $_POST['pesonCode'],
-        'aNumber' => 0,
-        'balance' => 0
-        ];
+        if (App::creatControll()) {
+            $array = [
+            'id' => rand(1000000000, 9999999999),
+            'name' => $_POST['firstName'],
+            'lastName' =>$_POST['lastName'],
+            'personCode' => $_POST['pesonCode'],
+            'aNumber' => App::acountNumber(),
+            'balance' => 0
+            ];
 
-        $this->get()->create($array);
+            $this->get()->create($array); 
+            App::addMassage('success', 'New account was successfully created.');
+        }else {
+            App::addMassage('danger', 'First name and last name must be longer then 3 symbols, first and last name must not have spaces and numbers, check personal code.');
+        }
+        
         App::redirect('creat');
     }
 
     public function delete($id) {
-        $this->get()->delete($id);
+        $account = $this->get()->show((int)$id);
+
+        if ($account['balance']== 0) {
+           $this->get()->delete($id); 
+            App::addMassage('success', 'Account was successfully deleted.');
+        }else {
+            App::addMassage('danger', 'You cannot delete this account.');
+        }
+    
         App::redirect('list');
+    }
+
+    public function update($id, $action) {
+     
+        $account = $this->get()->show((int)$id);
+        if ($_POST['money'] != '' && is_numeric($_POST['money'])) {
+            if ($action == 'add') {
+                $account['balance'] += $_POST['money'];
+                App::addMassage('success', 'Money was successfully added.');
+            }
+            elseif ($action == 'sub') {
+
+                if ($_POST['money'] <= $account['balance']) {
+                    $account['balance'] -= $_POST['money'];
+                    App::addMassage('success', 'Money was successfully sended.');
+                }else {
+                    App::addMassage('danger', "Please use correct input form use number.");
+                }
+            }
+            $this->get()->update($id, $account);
+        }else {
+            App::addMassage('danger', "Please use correct input form use number.");
+        }
+
+        $this->add($id);
     }
 
 }
